+++ cc_rig/templates/__init__.py
@@ -22,6 +22,7 @@ class FrameworkContent(TypedDict, total=False):
 
 _FRAMEWORK_MODULE_OVERRIDES: dict[str, str] = {
     "clap": "rust_cli",
+    "axum": "rust_web",
 }